+++ src/cli/commands.rs
@@ -1,4 +1,5 @@
 use clap::Subcommand;
+
 #[derive(Subcommand)]
 pub enum Commands {
     #[command(about = "Clone a repository into the specified directory.")]
@@ -28,6 +29,13 @@ pub enum Commands {
 
         #[clap(flatten)]
         regex_args: Regex,
+
+        #[arg(
+            long,
+            default_value_t = false,
+            help = "Only make a dry run and list the commands which would be executed."
+        )]
+        dry_run: bool,
     },
     #[command(about = "Make git add, git commit, git push in one go.")]
     Quick {
@@ -51,6 +59,13 @@ pub enum Commands {
             help = "Don't ask for permission to execute command per repository."
         )]
         skip_interactive: bool,
+        
+        #[arg(
+            long,
+            default_value_t = false,
+            help = "Only make a dry run and list the commands which would be executed."
+        )]
+        dry_run: bool,
     },
     #[command(about = "Execute a git command on a mass of repos.")]
     Mass {
@@ -71,13 +86,20 @@ pub enum Commands {
             help = "Don't ask for permission to execute command per repository."
         )]
         skip_interactive: bool,
+
+        #[arg(
+            long,
+            default_value_t = false,
+            help = "Only make a dry run and list the commands which would be executed."
+        )]
+        dry_run: bool,
     },
     #[command(about = "Manage your grgry profiles for different providers like github and gitlab.")]
     Profile {
         #[clap(subcommand)]
         sub: ProfileCommands,
     },
-    Test
+    // Test
 }
 
 #[derive(Subcommand)]
@@ -112,11 +134,11 @@ impl Regex {
     pub fn get_regex_args(&self, default: &str) -> (String, bool) {
         // Determine the value and whether reverse is true
         if let Some(rev_regex) = &self.rev_regex {
-            (rev_regex.clone(), true)
+            return (rev_regex.clone(), true)
         } else if let Some(regex) = &self.regex {
-            (regex.clone(), false);
+            return (regex.clone(), false);
         } else {
-            (default.to_string(), false)
+            return (default.to_string(), false)
         }
     }
 }
