+++ src/cli/mod.rs
@@ -1,2 +1,9 @@
-// mod commands;
-// mod main;
+pub mod clone;
+pub mod commands;
+pub mod profile;
+pub mod mass;
+pub mod quick;
+
+pub use clone::clone as clone;
+pub use mass::mass as mass;
+pub use quick::quick as quick;
