+++ src/config/config.rs
@@ -32,10 +32,7 @@ impl Config {
             Self::create_empty_config_file(&config_file_path);
         }
         let profiles: HashMap<String, Profile> = Self::load_profiles(&config_file_path);
-        Config {
-            profiles,
-            config_file_path,
-        }
+        Config { profiles, config_file_path }
     }
 
     fn get_default_config_path() -> PathBuf {
@@ -58,9 +55,7 @@ impl Config {
 
     fn load_profiles<P: AsRef<Path>>(config_file_path: P) -> HashMap<String, Profile> {
         let toml_content: String = fs::read_to_string(config_file_path).expect("Failed to read file");
-        let doc: DocumentMut = toml_content
-            .parse::<DocumentMut>()
-            .expect("Failed to parse TOML");
+        let doc: DocumentMut = toml_content.parse::<DocumentMut>().expect("Failed to parse TOML");
 
         //TODO on load check if not empty (VALID)
         let mut profiles: HashMap<String, Profile> = HashMap::new();
@@ -96,11 +91,8 @@ impl Config {
 
     // Saves the current state of profiles back to the config file
     pub fn save_config(&self) {
-        let mut toml_content: String =
-            fs::read_to_string(&self.config_file_path).expect("Failed to read file");
-        let mut doc: DocumentMut = toml_content
-            .parse::<DocumentMut>()
-            .expect("Failed to parse TOML");
+        let toml_content: String = fs::read_to_string(&self.config_file_path).expect("Failed to read file");
+        let mut doc: DocumentMut = toml_content.parse::<DocumentMut>().expect("Failed to parse TOML");
 
         // Remove profiles that are no longer in the HashMap
         let existing_keys: Vec<String> = doc.iter().map(|(key, _)| key.to_string()).collect();
@@ -143,7 +135,10 @@ impl Config {
         match self.profiles.values().find(|profile: &&Profile| profile.active) {
             Some(profile) => profile,
             None => {
-                eprintln!("{}", "One profile needs to be activated. For activating a profile use grgry profile activate!".red());
+                eprintln!(
+                    "{}",
+                    "One profile needs to be activated. For activating a profile use grgry profile activate!".red()
+                );
                 std::process::exit(1);
             }
         }
