+++ src/git_api/github.rs
@@ -4,8 +4,8 @@ use serde::Deserialize;
 use tokio::task::block_in_place;
 
 use crate::{
-    profile::config::Profile,
     git_api::git_providers::{call_api, get_repos_paralell, GitProvider, Repo},
+    config::config::Profile,
 };
 const PER_PAGE: i16 = 100;
 
@@ -33,18 +33,13 @@ impl Repo for GithubRepo {
 pub struct Github;
 impl GitProvider for Github {
     fn get_repos(
-        &self,
-        pat: &Option<String>,
-        collection_name: &str,
-        user: bool,
-        active_profile: Profile,
+        &self, pat: &Option<String>, collection_name: &str, user: bool, active_profile: Profile,
     ) -> Vec<Box<dyn Repo>> {
         block_in_place(|| {
             let future = async {
                 let collection_searchstring: &str = match user {
                     true => {
-                        if active_profile.username == collection_name && active_profile.token != ""
-                        {
+                        if active_profile.username == collection_name && active_profile.token != "" {
                             "user"
                         } else {
                             &format!("users/{}", collection_name)
@@ -52,10 +47,7 @@ impl GitProvider for Github {
                     }
                     false => &format!("orgs/{}", collection_name),
                 };
-                let endpoint: String = format!(
-                    "{}/{}/repos",
-                    &active_profile.baseaddress, collection_searchstring
-                ); //here the replace / --> %2F is not done because Github projects are top level on org or on user
+                let endpoint: String = format!("{}/{}/repos", &active_profile.baseaddress, collection_searchstring); //here the replace / --> %2F is not done because Github projects are top level on org or on user
                 let headers: Option<Vec<(String, String)>> = match pat {
                     Some(token) => Some(vec![
                         ("Authorization".to_string(), token.clone()),
@@ -66,14 +58,7 @@ impl GitProvider for Github {
                 let pages: i32 = self.get_page_number(&endpoint, headers.clone());
                 let parameters: Option<Vec<(String, String)>> =
                     Some(vec![("per_page".to_string(), PER_PAGE.to_string())]);
-                get_repos_paralell(
-                    pages,
-                    &endpoint,
-                    parameters,
-                    headers,
-                    &active_profile.provider,
-                )
-                .await
+                get_repos_paralell(pages, &endpoint, parameters, headers, &active_profile.provider).await
             };
 
             // Block on the async task, so it runs to completion and returns the result.
@@ -89,8 +74,7 @@ impl GitProvider for Github {
                     ("page".to_string(), "1".to_string()),
                     ("per_page".to_string(), PER_PAGE.to_string()),
                 ]);
-                let resp_total_repos: Response =
-                    call_api(endpoint, parameters.as_deref(), headers.as_deref()).await;
+                let resp_total_repos: Response = call_api(endpoint, parameters.as_deref(), headers.as_deref()).await;
                 let pages: i32 = match resp_total_repos.headers().get("link") {
                     Some(page) => {
                         let re: Regex = Regex::new(r"page=(\d+)").unwrap();
