+++ src/git_api/mod.rs
@@ -1,3 +1,3 @@
 pub mod git_providers;
-pub mod github;
-pub mod gitlab;
+mod github;
+mod gitlab;
