+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod git_api;
-pub mod profile;
+pub mod config;
 pub mod utils;
+pub mod cli;
